+++ CHANGES.md
@@ -11,6 +11,9 @@
 
 ## 2025.3
 
+- [UPDATE] デバイス権限処理を PermissionDispacher から Activity Result API へ移行
+  - PermissionDispacher が依存していた kapt も不要となったため依存ライブラリから削除
+  - @t-miya
 - [UPDATE] Kotlin バージョンを 2.0.20 に上げる
   - @t-miya
 - [UPDATE] Sora Android SDK を 2025.3.0 に上げる